pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-worklog"
version = "0.1.0"
description = "Automatic work time tracking with screen lock detection"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "dbus-python",
    "toggl-cli",
    "pendulum",
    "sortedcontainers",
    "PyGObject",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "mypy",
]

[project.scripts]
auto-worklog = "auto_worklog.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
